# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audaz LifeHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserNotifPrefs = {
  push?: boolean;
  whatsapp?: boolean;
};

export interface User {
  id: string;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  timezone?: string;
  notifPrefs?: UserNotifPrefs;
  createdAt?: string;
}

export type UserUpdateNotifPrefs = {
  push?: boolean;
  whatsapp?: boolean;
};

export interface UserUpdate {
  /** @nullable */
  whatsappNumber?: string | null;
  timezone?: string;
  notifPrefs?: UserUpdateNotifPrefs;
}

export interface Sector {
  id: string;
  label: string;
  color: string;
  isDefault: boolean;
}

export interface SectorInput {
  label: string;
  color?: string;
}

export interface SectorUpdate {
  label?: string;
  color?: string;
}

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  /** @nullable */
  deadline?: string | null;
  status: string;
  taskCount?: number;
  doneCount?: number;
  createdAt: string;
}

export interface Task {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  priority: number;
  status: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  reminderAt?: string | null;
  reminderSent?: boolean;
  reminderChannels?: string[];
  /** @nullable */
  sectorId?: string | null;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  goalId?: string | null;
  sector?: Sector | null;
  project?: Project | null;
  totalTimeSeconds?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface Goal {
  id: string;
  title: string;
  type: string;
  targetValue: number;
  currentValue: number;
  /** @nullable */
  deadline?: string | null;
  status: string;
  progressPercent?: number;
  isAtRisk?: boolean;
  createdAt: string;
}

export interface TimeSession {
  id: string;
  taskId: string;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  createdAt: string;
}

export interface TaskDetail {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  priority: number;
  status: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  reminderAt?: string | null;
  reminderSent?: boolean;
  reminderChannels?: string[];
  /** @nullable */
  sectorId?: string | null;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  goalId?: string | null;
  sector?: Sector | null;
  project?: Project | null;
  goal?: Goal | null;
  timeSessions: TimeSession[];
  totalTimeSeconds?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface TaskInput {
  title: string;
  description?: string;
  priority?: number;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  reminderAt?: string | null;
  reminderChannels?: string[];
  /** @nullable */
  sectorId?: string | null;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  goalId?: string | null;
}

export interface TaskUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  priority?: number;
  status?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  reminderAt?: string | null;
  reminderChannels?: string[];
  /** @nullable */
  sectorId?: string | null;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  goalId?: string | null;
}

export interface TaskSessionResult {
  task: Task;
  session: TimeSession;
}

export interface ProjectInput {
  name: string;
  /** @nullable */
  description?: string | null;
  color?: string;
  /** @nullable */
  deadline?: string | null;
}

export interface ProjectUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  color?: string;
  /** @nullable */
  deadline?: string | null;
  status?: string;
}

export interface GoalInput {
  title: string;
  type?: string;
  targetValue?: number;
  currentValue?: number;
  /** @nullable */
  deadline?: string | null;
}

export interface GoalUpdate {
  title?: string;
  type?: string;
  targetValue?: number;
  currentValue?: number;
  /** @nullable */
  deadline?: string | null;
  status?: string;
}

export interface Milestone {
  id: string;
  title: string;
  /** @nullable */
  dueDate?: string | null;
  done: boolean;
  /** @nullable */
  projectId?: string | null;
  createdAt: string;
}

export interface MilestoneInput {
  title: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  projectId?: string | null;
}

export interface MilestoneUpdate {
  title?: string;
  /** @nullable */
  dueDate?: string | null;
  done?: boolean;
  /** @nullable */
  projectId?: string | null;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface WeeklyStats {
  total: number;
  done: number;
  pending: number;
  executing: number;
  totalHours: number;
  byDay: DailyCount[];
}

export interface AnnualStats {
  heatmap: DailyCount[];
}

export type GetTasksParams = {
status?: string;
sectorId?: string;
priority?: number;
projectId?: string;
goalId?: string;
dateFrom?: string;
dateTo?: string;
};