# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Audaz LifeHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sectors:
    get:
      operationId: getSectors
      tags: [sectors]
      summary: List user sectors
      responses:
        "200":
          description: List of sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
    post:
      operationId: createSector
      tags: [sectors]
      summary: Create a sector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorInput"
      responses:
        "201":
          description: Created sector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /sectors/{id}:
    patch:
      operationId: updateSector
      tags: [sectors]
      summary: Update a sector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorUpdate"
      responses:
        "200":
          description: Updated sector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
    delete:
      operationId: deleteSector
      tags: [sectors]
      summary: Delete a sector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal (including currentValue increment)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /milestones:
    get:
      operationId: getMilestones
      tags: [milestones]
      summary: List milestones
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
...
components:
  schemas:
...
    Sector:
      type: object
      required: [id, label, color, isDefault]
      properties:
        id:
          type: string
        label:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
    SectorInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
        color:
          type: string
    SectorUpdate:
      type: object
      properties:
        label:
          type: string
        color:
          type: string
...
    Project:
      type: object
      required: [id, name, color, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        deadline:
          type: ["string", "null"]
        status:
          type: string
        taskCount:
          type: integer
        doneCount:
          type: integer
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        deadline:
          type: ["string", "null"]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        deadline:
          type: ["string", "null"]
        status:
          type: string
    Goal:
      type: object
      required: [id, title, type, targetValue, currentValue, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        deadline:
          type: ["string", "null"]
        status:
          type: string
        progressPercent:
          type: number
        isAtRisk:
          type: boolean
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        type:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        deadline:
          type: ["string", "null"]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        deadline:
          type: ["string", "null"]
        status:
          type: string
    Milestone:
      type: object
      required: [id, title, done, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        dueDate:
          type: ["string", "null"]
        done:
          type: boolean
        projectId:
          type: ["string", "null"]
        createdAt:
          type: string
    MilestoneInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        dueDate:
          type: ["string", "null"]
        projectId:
          type: ["string", "null"]
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: ["string", "null"]
        done:
          type: boolean
        projectId:
          type: ["string", "null"]
...