# lib/api-zod/src/generated/types/milestoneUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audaz LifeHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MilestoneUpdate {
  title?: string;
  /** @nullable */
  dueDate?: string | null;
  done?: boolean;
  /** @nullable */
  projectId?: string | null;
}