# lib/api-zod/src/generated/types/sector.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audaz LifeHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Sector {
  id: string;
  label: string;
  color: string;
  isDefault: boolean;
}